refactor: deleted unused files and transferred checkPermission method in uploadMiddleware.ts to authenticationMiddleware.ts
