--- src/config/auth/auth.ts
@@ -1,14 +1,0 @@
-export const ensureAuthenticated = (req, res, next) => {
-  if (req.isAuthenticated()) {
-    return next();
-  }
-  req.flash('error_msg', 'Please log in to view that resource');
-  res.redirect('/users/login');
-};
-
-export const forwardAuthenticated = (req, res, next) => {
-  if (!req.isAuthenticated()) {
-    return next();
-  }
-  res.redirect('/dashboard');
-};
