--- src/constants/application.ts
@@ -1,7 +1,0 @@
-const base = '/api';
-
-export default {
-  url: {
-    base,
-  },
-};
