+++ src/constants/constants.ts
@@ -3,3 +3,4 @@ export const ALLOWED_DOCUMENT_FILE_TYPES = ['.docx', '.pdf', '.doc'];
 export const ALLOWED_IMG_FILE_TYPES = ['.jpeg', '.jpg', '.png'];
 export const BACKEND_API = 'http://localhost:5000/api';
 export const RESET_PASSWORD_URL = 'http://localhost:3000/auth';
+export const BASE = '/api';
