+++ src/routes/user.route.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 import Utility from '../constants/utility';
 import { UserController } from '../controllers/user.controller';
-import { requireAdmin } from '../middlewares/userTypeHandler';
+import { requireAdmin } from '../middlewares/authenticationMiddleware';
 import user from './schema/user.schema';
 
 const passport = require('passport');
